replace thread-mill Shape with more robust model
Prevent solver dumping arcane coding errors to report window if  user  (legitimately)  sets "crest" value  to zero. Single tooth thread-mill tools exist which have a sharp point, ie zero flat "crest " surface. This should not provoke an error condition and should not present the machinist user with arcane developer output. This is fixed here. This model is a drop-in replacement for existing tool shape and covers both variants of this tool form: pointed and truncated. 

The problem is that setting a zero dimension makes two point concurrent and removes an edge from the model . This leads to and ill-defined model which crashes the solver. The user should not be exposed to this kind of coding breakage when entering legitimate data for a tool shape. The solution (or workaround) here is to add a very small faction of the shaft diameter to the user supplied zero to prevent the solver crashing. Since this is proportional to the tool shaft it will scale to any use of the tool shape without disrupting path accuracy.

Replaces accidentally closed #20258